+++ pages/api/v2/logout.ts
@@ -6,12 +6,8 @@ import withAuthentication from "../../../src/withAuthentication";
 async function handler(req: NextApiRequest, res: NextApiResponse, user: accounts) {
     if (req.method !== "GET") return res.status(405).json({ message: "Method not allowed" });
 
-    await prisma.sessions.deleteMany({
-        where: {
-            accountId: user.id,
-        }
-    });
-
+    // Delete all sessions for the user
+    await prisma.sessions.deleteMany({ where: { accountId: user.id, } });
     return res.status(200).json({ success: true, message: "Logged out" });
 }
